+++ src/application/globals/orderCollectionRequests.test.ts
@@ -2,7 +2,12 @@ import { orderCollectionRequests } from '../../application'
 
 describe('orderCollectionRequests()', () => {
   it('should order the postman request items in the order like defined', () => {
-    const order = ['POST::/crm/{id}', 'POST::/crm/monkies/{id}', 'GET::/crm/monkies/{id}']
+    const order = [
+      'POST::/crm/{id}',
+      'POST::/crm/monkies/{id}',
+      'GET::/crm/{crmId}/monkies/{id}',
+      'GET::/crm/monkies/{id}'
+    ]
     const obj = {
       item: [
         {
@@ -13,6 +18,30 @@ describe('orderCollectionRequests()', () => {
                 url: {
                   path: ['crm', 'monkies', ':id'],
                   variable: [
+                    {
+                      disabled: false,
+                      type: 'any',
+                      value: '<string>',
+                      key: 'id',
+                      description: '(Required) ID of the monkey you are acting upon.'
+                    }
+                  ]
+                },
+                method: 'GET'
+              }
+            },
+            {
+              request: {
+                url: {
+                  path: ['crm', ':crmId', 'monkies', ':id'],
+                  variable: [
+                    {
+                      disabled: false,
+                      type: 'any',
+                      value: '<string>',
+                      key: 'crmId',
+                      description: '(Required) ID of the crm you are acting upon.'
+                    },
                     {
                       disabled: false,
                       type: 'any',
@@ -157,6 +186,130 @@ describe('orderCollectionRequests()', () => {
                   ]
                 },
                 method: 'POST'
+              }
+            }
+          ]
+        }
+      ]
+    }
+
+    const transform = orderCollectionRequests(obj, order)
+    expect(transform).toMatchSnapshot()
+  })
+  it('should order the postman request items in the wildcard order', () => {
+    const order = ['POST::/*', 'GET::/*', 'PUT::/*']
+    const obj = {
+      item: [
+        {
+          name: 'Monkeys',
+          item: [
+            {
+              request: {
+                url: {
+                  path: ['crm', 'monkies', ':id'],
+                  variable: [
+                    {
+                      disabled: false,
+                      type: 'any',
+                      value: '<string>',
+                      key: 'id',
+                      description: '(Required) ID of the monkey you are acting upon.'
+                    }
+                  ]
+                },
+                method: 'GET'
+              }
+            },
+            {
+              request: {
+                url: {
+                  path: ['crm', ':crmId', 'monkies', ':id'],
+                  variable: [
+                    {
+                      disabled: false,
+                      type: 'any',
+                      value: '<string>',
+                      key: 'crmId',
+                      description: '(Required) ID of the crm you are acting upon.'
+                    },
+                    {
+                      disabled: false,
+                      type: 'any',
+                      value: '<string>',
+                      key: 'id',
+                      description: '(Required) ID of the monkey you are acting upon.'
+                    }
+                  ]
+                },
+                method: 'GET'
+              }
+            },
+            {
+              request: {
+                url: {
+                  path: ['crm', 'monkies', ':id'],
+                  variable: [
+                    {
+                      disabled: false,
+                      type: 'any',
+                      value: '<string>',
+                      key: 'id',
+                      description: '(Required) ID of the monkey you are acting upon.'
+                    }
+                  ]
+                },
+                method: 'POST'
+              }
+            },
+            {
+              request: {
+                url: {
+                  path: ['crm', ':id'],
+                  variable: [
+                    {
+                      disabled: false,
+                      type: 'any',
+                      value: '<string>',
+                      key: 'id',
+                      description: '(Required) ID of the monkey you are acting upon.'
+                    }
+                  ]
+                },
+                method: 'GET'
+              }
+            },
+            {
+              request: {
+                url: {
+                  path: ['crm', ':id'],
+                  variable: [
+                    {
+                      disabled: false,
+                      type: 'any',
+                      value: '<string>',
+                      key: 'id',
+                      description: '(Required) ID of the monkey you are acting upon.'
+                    }
+                  ]
+                },
+                method: 'POST'
+              }
+            },
+            {
+              request: {
+                url: {
+                  path: ['crm', ':id'],
+                  variable: [
+                    {
+                      disabled: false,
+                      type: 'any',
+                      value: '<string>',
+                      key: 'id',
+                      description: '(Required) ID of the monkey you are acting upon.'
+                    }
+                  ]
+                },
+                method: 'PUT'
               }
             }
           ]