+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - ContentTest - Check maximum length of array or string of response body
 - ContentTest - Check minimum length of array or string of response header
 - ContentTest - Check maximum length of array or string of response header
+- Fix for orderOfOperations for operations have multiple path variables (#208)
+- Implemented suggested fix for orderOfOperations with * wildcard (#122)
 
 ## v1.12.2 - (2021-12-15)